feat: add portable YAML split and Node-only subentry
- Add src/utils/yaml-shared.ts: portable YAML parse/validate/export with no fs dependency
- Slim src/utils/yaml-loader.ts: thin Node wrapper delegating to yaml-shared
- Add src/node/scanner.ts: NodeScanner extends Scanner with customPatternFiles support
- Add src/node/index.ts: ./node entrypoint re-exporting all portable API plus Node-only exports
- Add ./node export to package.json exports map with provenance publish config
- Add protected reloadPatterns/getAdditionalPatterns hooks to base Scanner
- Remove customPatternFiles from ScannerConfig (now lives on NodeScannerConfig)
- Remove duplicate GCP/Azure patterns from aws.ts (now owned by their dedicated files)
- Fix stale pattern-count tests to account for gcp/azure/chat-platform collections
- All 144 tests pass
