+++ package.json
@@ -3,16 +3,25 @@
   "version": "1.1.1",
   "description": "TypeScript DLP scanning library for detecting secrets and sensitive data",
   "type": "module",
+  "publishConfig": {
+    "access": "public",
+    "provenance": true
+  },
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
+    },
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "import": "./dist/node/index.js"
     }
   },
   "files": [
     "dist",
+    "patterns.example.yaml",
     "README.md",
     "LICENSE"
   ],
