+++ src/node/index.ts
@@ -1,0 +1,7 @@
+export * from '../index.js'
+export { NodeScanner, type NodeScannerConfig } from './scanner.js'
+export {
+  loadPatternsFromYAML,
+  loadPatternsFromYAMLString,
+  exportPatternsToYAML,
+} from '../utils/yaml-loader.js'
