+++ src/patterns/aws.ts
@@ -73,100 +73,4 @@ export const awsPatterns: PatternDefinition[] = [
       negative: [],
     },
   },
-
-  // GCP Patterns
-  {
-    id: 'gcp-api-key',
-    name: 'Google Cloud API Key',
-    description: 'Google Cloud Platform API key',
-    severity: 'high',
-    pattern: /AIza[0-9A-Za-z\-_]{35}/g,
-    keywords: ['google', 'gcp', 'api', 'key', 'cloud'],
-    validators: [{ type: 'entropy', min: 4.0 }],
-    examples: {
-      positive: ['AIzaSyDaGmWKa4JsXZ-HjGw7ISLn_3namBGewQe'],
-      negative: [],
-    },
-  },
-  {
-    id: 'gcp-service-account',
-    name: 'Google Cloud Service Account',
-    description: 'GCP service account email pattern',
-    severity: 'medium',
-    pattern: /[a-z0-9-]+@[a-z0-9-]+\.iam\.gserviceaccount\.com/gi,
-    keywords: ['service', 'account', 'gcp', 'google', 'iam'],
-    validators: [],
-    examples: {
-      positive: ['my-service@my-project.iam.gserviceaccount.com'],
-      negative: [],
-    },
-  },
-  {
-    id: 'gcp-oauth-token',
-    name: 'Google OAuth Token',
-    description: 'Google OAuth access token',
-    severity: 'high',
-    pattern: /ya29\.[0-9A-Za-z\-_]+/g,
-    keywords: ['google', 'oauth', 'token', 'access'],
-    validators: [{ type: 'entropy', min: 4.0 }],
-    examples: {
-      positive: ['ya29.a0AfB_byC1234567890abcdefghijklmnop'],
-      negative: [],
-    },
-  },
-
-  // Azure Patterns
-  {
-    id: 'azure-storage-key',
-    name: 'Azure Storage Account Key',
-    description: 'Microsoft Azure storage account access key',
-    severity: 'critical',
-    pattern: /(?<![A-Za-z0-9/+=])[A-Za-z0-9/+=]{86}==(?![A-Za-z0-9/+=])/g,
-    keywords: ['azure', 'storage', 'account', 'key', 'microsoft'],
-    validators: [{ type: 'entropy', min: 5.0 }],
-    examples: {
-      positive: [],
-      negative: [],
-    },
-  },
-  {
-    id: 'azure-connection-string',
-    name: 'Azure Connection String',
-    description: 'Azure service connection string',
-    severity: 'critical',
-    pattern: /DefaultEndpointsProtocol=https?;AccountName=[^;]+;AccountKey=[A-Za-z0-9/+=]+;?/gi,
-    keywords: ['azure', 'connection', 'string', 'endpoint'],
-    validators: [],
-    examples: {
-      positive: ['DefaultEndpointsProtocol=https;AccountName=myaccount;AccountKey=abc123=='],
-      negative: [],
-    },
-  },
-  {
-    id: 'azure-client-secret',
-    name: 'Azure AD Client Secret',
-    description: 'Azure Active Directory application client secret',
-    severity: 'critical',
-    pattern: /[a-zA-Z0-9~_.-]{34,40}/g,
-    keywords: ['azure', 'client', 'secret', 'ad', 'active', 'directory'],
-    validators: [{ type: 'entropy', min: 4.5 }],
-    falsePositives: [/^[a-z]+$/i, /^[0-9]+$/],
-    examples: {
-      positive: [],
-      negative: [],
-    },
-  },
-  {
-    id: 'azure-sas-token',
-    name: 'Azure SAS Token',
-    description: 'Azure Shared Access Signature token',
-    severity: 'high',
-    pattern: /[?&]sig=[A-Za-z0-9%]+(&|$)/g,
-    keywords: ['azure', 'sas', 'signature', 'token', 'blob'],
-    validators: [],
-    examples: {
-      positive: ['?sv=2020-08-04&sig=abc123%3D'],
-      negative: [],
-    },
-  },
 ]
