+++ src/types.ts
@@ -24,7 +24,6 @@ export interface ScannerConfig {
   patterns?: string[]
   exclude?: string[]
   customPatterns?: PatternDefinition[]
-  customPatternFiles?: string[] // YAML files to load patterns from
   minConfidence?: number
   includeContext?: boolean
   contextLines?: number
